chore: Push stdout/stderr retrieval down from container-level
std.io.getStdOut() on Windows is trying to call external functions,
which is not allowed at comptime. Running this function in the top-level
of zli.zig - i.e. at "container-level" - will prevent building. On a
side-note on Linux this is not failing, because there the function call
is returning the default POSIX STDOUT_FILENO.

This commit moves the top-level stdout/stderr retrieval into the Command
struct as well as the helper-functions, which are not run at
compile-time.